libxc/xc_sr: factor out filter_pages()
When processing a PAGE_DATA record, the restore side needs to set the
types of incoming pages using the appropriate restore op and filter the
list of pfns in the record to the subset that are 'backed' - ie.
accompanied by real backing data in the stream that needs to be filled
in.

Both of these steps are also required when processing postcopy records,
so factor it out into a common helper.

No functional change.

Signed-off-by: Joshua Otto <xxx>