libxl/migration: add precopy tuning parameters
In the context of the live migration algorithm, the precopy iteration
count refers to the number of page-copying iterations performed prior to
the suspension of the guest and transmission of the final set of dirty
pages.  Similarly, the precopy dirty threshold refers to the dirty page
count below which we judge it more profitable to proceed to
stop-and-copy rather than continue with the precopy.  These would be
helpful tuning parameters to work with when migrating particularly busy
guests, as they enable an administrator to reap the available benefits
of the precopy algorithm (the transmission of guest pages _not_ in the
writable working set can be completed without guest downtime) while
reducing the total amount of time required for the migration (as
iterations of the precopy loop that will certainly be redundant can be
skipped in favour of an earlier suspension).

To expose these tuning parameters to users:
- introduce a new libxl API function, libxl_domain_live_migrate(),
  taking the same parameters as libxl_domain_suspend() _and_
  precopy_iterations and precopy_dirty_threshold parameters, and
  consider these parameters in the precopy policy

  (though a pair of new parameters on their own might not warrant an
  entirely new API function, it is added in anticipation of a number of
  additional migration-only parameters that would be cumbersome on the
  whole to tack on to the existing suspend API)

- switch xl migrate to the new libxl_domain_live_migrate() and add new
  --postcopy-iterations and --postcopy-threshold parameters to pass
  through

Signed-off-by: Joshua Otto <xxx>